% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASAP.get.depth.parallel.R
\name{ASAP.get.depth.parallel}
\alias{ASAP.get.depth.parallel}
\title{Process ASAP DF for Depth Data in parallel}
\usage{
ASAP.get.depth.parallel(read.ASAP.df, num_cores = 1)
}
\arguments{
\item{read.ASAP.df}{imported df}

\item{num_cores}{number of cores to use for parallel processing}
}
\value{
An object with data from the read.ASAP dataframe.
}
\description{
Function will use a user specified dataframe from read.ASAP function and parse the depth data into a df that can be easily plotted.
}
