% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASAP.get.depth.R
\name{ASAP.get.depth}
\alias{ASAP.get.depth}
\title{Process ASAP XML List for Depth Data}
\usage{
ASAP.get.depth(FinalList)
}
\arguments{
\item{List}{of all XML objects to extract depth data from.}
}
\value{
An object with data from the ASAP xml.
}
\description{
Process ASAP XML List for Depth Data
}
