% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASAP.get.metrics.R
\name{ASAP.get.metrics}
\alias{ASAP.get.metrics}
\title{Process ASAP XML List for General Metrics}
\usage{
ASAP.get.metrics(FinalList)
}
\arguments{
\item{List}{of all XML objects to extract depth data from.}
}
\value{
An object with general metrics by sample.
}
\description{
Process ASAP XML List for General Metrics
}
