% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASAP.get.proportions.R,
%   R/ASAP.get.proportions.parallel.R
\name{ASAP.get.proportions}
\alias{ASAP.get.proportions}
\title{Process ASAP DF for Proportions Data}
\usage{
ASAP.get.proportions(read.ASAP.df, num_cores = 1)

ASAP.get.proportions(read.ASAP.df, num_cores = 1)
}
\arguments{
\item{read.ASAP.df}{imported df}

\item{num_cores}{number of cores to use for parallel processing}
}
\value{
An object with data from the read.ASAP dataframe.

An object with data from the read.ASAP dataframe.
}
\description{
Function will use a user specified dataframe from read.ASAP function and parse the proportions data into a df that can be easily plotted.

Function will use a user specified dataframe from read.ASAP function and parse the proportions data into a df that can be easily plotted.
}
