% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASAP.get.SMOR.discards.R,
%   R/ASAP.get.SMOR.discards.parallel.R
\name{ASAP.get.smor.discards}
\alias{ASAP.get.smor.discards}
\title{Process ASAP DF for SMOR discards}
\usage{
ASAP.get.smor.discards(read.ASAP.df, num_cores = 1)

ASAP.get.smor.discards(read.ASAP.df, num_cores = 1)
}
\arguments{
\item{read.ASAP.df}{imported df}

\item{num_cores}{number of cores to use for parallel processing}
}
\value{
An object with data from the read.ASAP dataframe.

An object with data from the read.ASAP dataframe.
}
\description{
Function will use a user specified dataframe from read.ASAP function and parse the SMOR discards data into a df that can be easily plotted.

Function will use a user specified dataframe from read.ASAP function and parse the SMOR discards data into a df that can be easily plotted.
}
