% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASAP.get.SNPS.R
\name{ASAP.get.SNPS}
\alias{ASAP.get.SNPS}
\title{Process ASAP XML List for SNPS}
\usage{
ASAP.get.SNPS(FinalList)
}
\arguments{
\item{List}{of all XML objects to extract SNP data from R parsed XML.}
}
\value{
An object with data regarding SNPS from the ASAP xml.
}
\description{
Process ASAP XML List for SNPS
}
