% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASAP_get_metrics.R
\name{ASAP_get_metrics}
\alias{ASAP_get_metrics}
\title{Process ASAP XML List for General Metrics}
\usage{
ASAP_get_metrics(FinalList)
}
\arguments{
\item{List}{of all XML objects to extract depth data from.}
}
\value{
An object with general metrics by sample.
}
\description{
Process ASAP XML List for General Metrics
}
