% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASAP_get_proportions.R
\name{ASAP_get_proportions}
\alias{ASAP_get_proportions}
\title{Process ASAP DF for Proportions Data}
\usage{
ASAP_get_proportions(read.ASAP.df)
}
\arguments{
\item{read.ASAP}{imported df}
}
\value{
An object with data from the read.ASAP dataframe.
}
\description{
Function will use a user specified dataframe from read.ASAP function and parse the proportions data into a df that can be easily plotted.
}
