% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASAP_get_SMOR_discards.R
\name{ASAP_get_smor_discards}
\alias{ASAP_get_smor_discards}
\title{Process ASAP DF for SMOR discards}
\usage{
ASAP_get_smor_discards(read.ASAP.df)
}
\arguments{
\item{read.ASAP}{imported df}
}
\value{
An object with data from the read.ASAP dataframe.
}
\description{
Function will use a user specified dataframe from read.ASAP function and parse the SMOR discards data into a df that can be easily plotted.
}
