% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASAP_get_SNPS.R
\name{ASAP_get_SNPS}
\alias{ASAP_get_SNPS}
\title{Process ASAP XML List for SNPS}
\usage{
ASAP_get_SNPS(FinalList)
}
\arguments{
\item{List}{of all XML objects to extract SNP data from R parsed XML.}
}
\value{
An object with data regarding SNPS from the ASAP xml.
}
\description{
Process ASAP XML List for SNPS
}
