% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASAP_plot_sars_depth.R
\name{ASAP_plot_sars_depth}
\alias{ASAP_plot_sars_depth}
\title{Plot SARS Coverage Depth}
\usage{
ASAP_plot_sars_depth(XML_List, XML_Names, Out_Dir)
}
\arguments{
\item{XML_List}{A list of XML path.}

\item{XML_Names}{A list of XML names.}

\item{Out_Dir}{A directory of where the output should be saved.}
}
\value{
Objects should be saved to specified directory.
}
\description{
This function creates plots to display the coverage across the genome of select ASAP XMLs.
}
