% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.bed.files.R
\name{create.asap.bed.file}
\alias{create.asap.bed.file}
\title{Create an ASAP formatted bed file}
\usage{
create.asap.bed.file(find.primers.df, save.path = NA)
}
\arguments{
\item{find.primers.df}{A dataframe from the find.primers function.}

\item{save.path}{A path to automatically save the bed file to in a .bed format.}
}
\value{
A dataframe formatted as a bed file.
}
\description{
This function creates a bed file for ASAP from find.primers output.
}
