% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.ASAP.fasta.R
\name{export.ASAP.fasta}
\alias{export.ASAP.fasta}
\title{Export ASAP Consensus Fasta}
\usage{
export.ASAP.fasta(read.ASAP.df, savepath)
}
\arguments{
\item{read.ASAP.df}{from read.ASAP with added column "Sequence_Name"}

\item{savepath}{path to save the file to.}
}
\value{
Statement of export completion
}
\description{
This function exports the consensus sequence from read.ASAP dataframe.
}
