% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.diff.in.seq.R
\name{find.diff.in.seq}
\alias{find.diff.in.seq}
\title{Find Differences between two sequence}
\usage{
find.diff.in.seq(a, b)
}
\arguments{
\item{a}{A character reference string.}

\item{b}{A character string that is compared to the reference}
}
\value{
A character string of differences within the sequence
}
\description{
Find differences between two sequences and return the differences separated by ",".
Note: This function needs to be adjusted to deal with IUPAC ambiguity
}
