% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.primers.R
\name{find.primers}
\alias{find.primers}
\title{Find primers in reference sequence}
\usage{
find.primers(
  fasta.path,
  primer.names,
  primer.direction,
  primer.list,
  max.mismatch
)
}
\arguments{
\item{fasta.path}{The path to the reference fasta file.}

\item{primer.names}{A list of primer names.}

\item{primer.direction}{A list of primer directions.}

\item{primer.list}{A list of primers to align to the reference file.}

\item{max.mismatch}{A number of mismatches to allow in search. Note: that the script records ambigous characters as mismatches, thus, ensure this number is greater than the number of ambigous characters.}
}
\value{
An dataframe with the reference name, primer name, primer direction, primer sequence, searched direction, start location, end location, matching width, and returned sequence.
}
\description{
This function uses reference sequence and a list of primers to identify and search for the locations, returning a dataframe with the reference name, primer name, primer direction, primer sequence, searched direction, start location, end location, matching width, and returned sequence. Be aware that this function searches both the original and reverse complimented sequences.
}
