% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.ASAP.individual.R
\name{read.ASAP.individual}
\alias{read.ASAP.individual}
\title{Read ASAP Results for an Individual xml file}
\usage{
read.ASAP.individual(XML)
}
\arguments{
\item{XML}{A path with the individual ASAP XML file to parse.}
}
\value{
An object with data from the ASAP xml.
}
\description{
This function uses a user provided ASAP results XML path to parse sample and assay information from the ASAP output for an individual sample. SNPs will not be included in this function and the function read.ASAP.snps should be used instead.
}
