% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.ASAP.R
\name{read.ASAP}
\alias{read.ASAP}
\title{Read ASAP Results}
\usage{
read.ASAP(XML)
}
\arguments{
\item{XML}{A path with the ASAP XML file to parse.}
}
\value{
An object with data from the ASAP xml.
}
\description{
This function uses a user provided ASAP results XML path to parse sample and assay information from the ASAP output. SNPs will not be included in this function and the function read.ASAP.snps should be used instead.
}
