% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.ASAP.snps.individual.R
\name{read.ASAP.snps.individual}
\alias{read.ASAP.snps.individual}
\title{Read ASAP SNP Results from a single XML}
\usage{
read.ASAP.snps.individual(XML)
}
\arguments{
\item{XML}{A path with the ASAP XML file to parse.}
}
\value{
An object with data from the ASAP xml.
}
\description{
This function uses a user provided individual ASAP results XML path to parse sample and assay information from the ASAP output and display the SNPs.
}
