% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.ASAP.snps.parallel.R
\name{read.ASAP.snps.parallel}
\alias{read.ASAP.snps.parallel}
\title{Read ASAP SNP Results in Parallel}
\usage{
read.ASAP.snps.parallel(XML, cores)
}
\arguments{
\item{XML}{A path with the ASAP XML file to parse.}
}
\value{
An object with data from the ASAP xml.
}
\description{
This function uses a user provided ASAP results XML path to parse sample and assay information from the ASAP output and display the SNPs.
}
