% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASAP.get.depth.R, R/ASAP_get_depth.R
\name{ASAP.get.depth}
\alias{ASAP.get.depth}
\title{Process ASAP DF for Depth Data}
\usage{
ASAP.get.depth(read.ASAP.df)

ASAP.get.depth(read.ASAP.df)
}
\arguments{
\item{read.ASAP}{imported df}
}
\value{
An object with data from the read.ASAP dataframe.

An object with data from the read.ASAP dataframe.
}
\description{
Function will use a user specified dataframe from read.ASAP function and parse the depth data into a df that can be easily plotted.

Function will use a user specified dataframe from read.ASAP function and parse the depth data into a df that can be easily plotted.
}
