% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASAP.read.sars.xml.R
\name{ASAP.read.sars.xml}
\alias{ASAP.read.sars.xml}
\title{Import ASAP XML Results}
\usage{
ASAP.read.sars.xml(XML_List, XML_Names)
}
\arguments{
\item{XML_List}{A list of XML path.}

\item{XML_Names}{A list of XML names.}
}
\value{
An object with data from the ASAP xml.
}
\description{
This function uses a user provided ASAP results XML path and a user provided XML name to read in the XML into a matrix. Several other functions can then be used to parse the matrix.
}
